backend/backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
beautifulsoup4==4.13.4
boto3==1.38.39
botocore==1.38.39
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
contourpy==1.3.2
Crawl4AI==0.6.3
cryptography==45.0.5
cssselect==1.3.0
cycler==0.12.1
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fasta2a==0.3.1
fastapi==0.110.0
fastavro==1.11.1
filelock==3.18.0
fonttools==4.58.1
frozenlist==1.7.0
fsspec==2025.5.1
fuzzywuzzy==0.18.0
google-auth==2.40.3
google-genai==1.21.1
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
groq==0.28.0
h11==0.16.0
hf-xet==1.1.4
html5lib==1.1
httpcore==1.0.9
httpx
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
Levenshtein==0.27.1
litellm==1.74.0.post1
logfire==3.21.1
logfire-api==3.21.1
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mcp==1.9.4
mcp-server-fetch
mdurl==0.1.2
mistralai==1.8.2
multidict==6.6.3
nltk==3.9.1
numpy==2.2.6
openai==1.88.0
openai-agents==0.0.19
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pandas==2.3.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==10.4.0
playwright==1.53.0
prompt_toolkit==3.0.51
propcache==0.3.2
Protego==0.5.0
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-ai==0.3.1
pydantic-ai-slim==0.3.1
pydantic-evals==0.3.1
pydantic-graph==0.3.1
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.3
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
readabilipy==0.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.36.3
tenacity==8.5.0
tf-playwright-stealth==1.2.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
xgboost==3.0.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
